Move app-specific settings to new lute/settings/current.py module.
